# frontend/openapi.json
{"openapi":"3.0.3","info":{"title":"Report Platform API","version":"1.0.0","description":"API for asynchronous report generation and retrieval."},"servers":[{"url":"http://localhost:3001","description":"Local development"}],"tags":[{"name":"Health"},{"name":"Report Templates"},{"name":"Reports"}],"components":{"schemas":{"RunStatus":{"type":"string","enum":["PENDING","PROCESSING","COMPLETED","FAILED"]},"ReportFormat":{"type":"string","enum":["PDF","XLSX"]},"ReportTemplate":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string","nullable":true},"format":{"$ref":"#/components/schemas/ReportFormat"},"type":{"type":"string","nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"Report":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"reportTemplateId":{"type":"string","format":"uuid"},"status":{"$ref":"#/components/schemas/RunStatus"},"startedAt":{"type":"string","format":"date-time","nullable":true},"completedAt":{"type":"string","format":"date-time","nullable":true},"error":{"type":"string","nullable":true},"resultUrl":{"type":"string","nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PaginatedReportsResponse":{"type":"object","properties":{"reports":{"type":"array","items":{"$ref":"#/components/schemas/Report"}},"pagination":{"type":"object","properties":{"page":{"type":"integer","minimum":1},"limit":{"type":"integer","minimum":1},"total":{"type":"integer","minimum":0},"totalPages":{"type":"integer","minimum":0}}}}},"StartGenerationResponse":{"type":"object","properties":{"message":{"type":"string"},"runId":{"type":"string","format":"uuid"},"statusUrl":{"type":"string"}}},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}}}},"paths":{"/health":{"get":{"tags":["Health"],"summary":"Health check","responses":{"200":{"description":"Service is healthy"}}}},"/api/report-templates":{"get":{"tags":["Report Templates"],"summary":"Get all report templates","responses":{"200":{"description":"List of report templates","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReportTemplate"}}}}}}}},"/api/report-templates/{id}":{"get":{"tags":["Report Templates"],"summary":"Get report template by id","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Report template","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportTemplate"}}}},"404":{"description":"Template not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/report-templates/{id}/generate":{"post":{"tags":["Report Templates"],"summary":"Start asynchronous report generation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"202":{"description":"Generation started","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartGenerationResponse"}}}},"404":{"description":"Template not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/reports":{"get":{"tags":["Reports"],"summary":"Get report runs with pagination","parameters":[{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"default":10}}],"responses":{"200":{"description":"Paginated report runs","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedReportsResponse"}}}}}}},"/api/reports/{id}/status":{"get":{"tags":["Reports"],"summary":"Get report run status","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Report run","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}},"404":{"description":"Report not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/reports/{id}/download":{"get":{"tags":["Reports"],"summary":"Download generated report","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Report file (PDF or XLSX)"},"404":{"description":"Result not available","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}}}